requirements.txt
einops==0.8.1
local-attention==1.11.2
reformer-pytorch==1.4.4

numpy==2.1.2
scipy==1.16.3
scikit-learn==1.7.2
pandas==2.3.3
matplotlib==3.10.8
sktime==0.40.1
sympy==1.13.1
PyWavelets==1.9.0
datasets==4.5.0

tqdm==4.66.5
patool==4.0.3

transformers==4.57.3
#datasets==4.4.1
huggingface_hub==0.36.0
chronos-forecasting==2.2.1
tirex-ts==1.3.0
timesfm==1.3.0

# uni2ts deps
######
gluonts==0.16.2
lightning==2.6.0
hydra-core==1.3.0
jax==0.8.1
jaxtyping==0.3.4
orjson==3.11.5
tensorboard==2.20.0
python-dotenv==1.0.0
######

